+++ packages/provider/src/providers/index.ts
@@ -1,6 +1,0 @@
-export * from './coinbase'
-export * from './metamask'
-export * from './near'
-export * from './near-raw-provider'
-export * from './phantom'
-export * from './solflare'