+++ packages/react-nft-checkout/vite.config.ts
@@ -40,7 +40,6 @@ export default defineConfig(({ mode }) => {
       dedupe: ['react'],
       alias: {
         '@': `${root}/`,
-        'near-api-js': 'near-api-js/dist/near-api-js.js',
         '@uniswap/router-sdk': '@uniswap/router-sdk/dist/router-sdk.esm.js',
       },
     },