+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `root`: `@rarimo/providers-evm` package
+- `root`: `@rarimo/providers-near` package
+- `root`: `@rarimo/providers-solana` package
+
+### Changed
+- `@rarimo/provider`: Wallet wrappers split to the several packages: `@rarimo/providers-evm`, `@rarimo/providers-near`, `@rarimo/providers-solana`
+
+### Fixed
+- `@rarimo/nft-checkout`: Uniswap V3 token list url changed to the new one to omit gateway issues
 
 ## [1.5.0-rc.8] - 2023-04-25
 ### Added
@@ -181,7 +191,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Under the hood changes
 - Initiated repo with `@rarimo/provider` and `@rarimo/nft-checkout` packages
 
-[Unreleased]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.7...HEAD
+[Unreleased]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.8...HEAD
+[1.5.0-rc.8]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.7...1.5.0-rc.8
 [1.5.0-rc.7]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.6...1.5.0-rc.7
 [1.5.0-rc.6]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.5...1.5.0-rc.6
 [1.5.0-rc.5]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.4...1.5.0-rc.5